Idea for monthly / annual data
Here's an idea I worked up this weekend that would allow monthly data to have "MM YYYY" labels; and annual data to have "YYYY" data, even if there are fewer than 8 years (or whatever tolerance level is defined).

Not really proposing this as a true pull. Just wanted to share the code with you in case you might want to play with it, extend it, etc.